modify buffer demo

+++ 06-buffer/05-bmp-write.js
@@ -18,7 +18,7 @@ buf.writeUInt32LE(totalSize, 0x02);
 buf.writeUInt32LE(54, 0x0a);
 buf.writeUInt32LE(40, 0x0e);
 buf.writeUInt16LE(1, 0x1a);
-buf.writeUInt32LE(32, 0x1c);
+buf.writeUInt16LE(32, 0x1c);
 buf.writeUInt32LE(pixelByteSize, 0x22);
 buf.writeInt32LE(width, 0x12);
 buf.writeInt32LE(height, 0x16);
